refactor(core): move the design tokens into the render-free core
Feature 017, phases 1-2. Groundwork for the Material 3 visual system: the
tokens move to a crate that cannot name a rendering type, and the two pieces
of UI decision logic that belong outside a widget follow them.

  * Tokens relocate from micold-client to micold-core. The module already
    imported only from the core, so no dependency edge was added. The move is
    strictly mechanical - same values, new home - guarded by tokens_move.rs,
    which asserts all 22 role values, 22 tag values and every scale constant
    are byte-identical. Re-valuing them is feature 018's work.

  * The WCAG AA contrast gate moves with them. It only ever used core types,
    so it now runs in a crate with no rendering dependency: "tokens are
    render-free" becomes a compile error rather than a convention.

  * Unified floating-surface dismissal lands as pure logic in core::overlay.
    The app grew five independent overlay implementations, each answering
    "when does this close?" for itself, so the answers drifted. Deciding
    whether a trigger dismisses a surface is branching logic, which
    Principle I requires to be testable and the GUI-wiring exception does not
    cover - so the widget will own presentation and this owns the rule. A
    totality test proves every surface/trigger combination is defined.

  * A resolved-style parity baseline is recorded as a committed fixture:
    116 styles across every style function, widget status and both schemes,
    asserted byte-for-byte. Feature 017 must change no appearance, and this
    catches drift exhaustively and names the component and status that moved.
    It does not see layout or widget-tree structure, so a reduced visual pass
    is still required before the wrapper phases.

30 of the 34 touched files change by one import line only.

Tests: 800 passing, 0 failing. Clippy clean. The parity snapshot confirms the
move altered no resolved style.

+++ crates/micold-core/src/tokens.rs
@@ -171,15 +171,15 @@ pub fn roles(scheme: ColorScheme) -> Roles {
 /// applied in the GUI via `iced::Font`.
 pub mod type_scale {
     /// Large empty-state / dialog headline.
-    pub const DISPLAY: u16 = 32;
+    pub const DISPLAY: f32 = 32.0;
     /// Active-project name, section headers.
-    pub const HEADLINE: u16 = 24;
+    pub const HEADLINE: f32 = 24.0;
     /// App-bar title, list-item primary text.
-    pub const TITLE: u16 = 18;
+    pub const TITLE: f32 = 18.0;
     /// Default body text, descriptions.
-    pub const BODY: u16 = 14;
+    pub const BODY: f32 = 14.0;
     /// Paths, captions, badges.
-    pub const LABEL: u16 = 12;
+    pub const LABEL: f32 = 12.0;
 }
 
 /// Sidebar-scoped type sizes — 80% of the app-wide scale (FR-012). Applied ONLY within the
@@ -187,30 +187,30 @@ pub mod type_scale {
 /// "80%" intent is auditable (`tests/tokens.rs`).
 pub mod sidebar {
     /// Worktree display name — 80% of `type_scale::BODY` (14 → 11).
-    pub const NAME: u16 = 11;
+    pub const NAME: f32 = 11.0;
     /// Tag chip text — 80% of `type_scale::LABEL` (12 → 10).
-    pub const TAG: u16 = 10;
+    pub const TAG: f32 = 10.0;
     /// Session label — 80% of `type_scale::BODY` (14 → 11).
-    pub const SESSION: u16 = 11;
+    pub const SESSION: f32 = 11.0;
 }
 
 /// Spacing scale, in logical pixels. All padding/gaps use these steps (SC-007).
 pub mod spacing {
-    pub const XS: u16 = 4;
-    pub const SM: u16 = 8;
-    pub const MD: u16 = 16;
-    pub const LG: u16 = 24;
-    pub const XL: u16 = 32;
+    pub const XS: f32 = 4.0;
+    pub const SM: f32 = 8.0;
+    pub const MD: f32 = 16.0;
+    pub const LG: f32 = 24.0;
+    pub const XL: f32 = 32.0;
 }
 
 /// Corner radii, in logical pixels.
 pub mod shape {
     /// Buttons, badges.
-    pub const SM: u16 = 8;
+    pub const SM: f32 = 8.0;
     /// Cards / list items / surfaces.
-    pub const MD: u16 = 12;
+    pub const MD: f32 = 12.0;
     /// Dialogs.
-    pub const LG: u16 = 16;
+    pub const LG: f32 = 16.0;
     /// Pills.
-    pub const FULL: u16 = 9999;
+    pub const FULL: f32 = 9999.0;
 }
